style(header): move navigation and header to layout folder
