+++ src/components/Layout/Header.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/Variables.scss";
+@import "../../styles/Variables";
 
 .header {
   display: flex;
