+++ src/components/Layout/Navigation.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/Variables.scss";
+@import "../../styles/Variables";
 
 .nav {
   display: flex;
