+++ src/components/Layout/Navigation.tsx
@@ -1,18 +1,18 @@
 import Link from "next/link"
 import style from "./Navigation.module.scss"
-import text from "../styles/Text.module.css"
-import effects from "../styles/Effects.module.scss"
+import text from "../../styles/Text.module.css"
+import effects from "../../styles/Effects.module.scss"
 import cs from "classnames"
-import { Button } from "./Button"
+import { Button } from "../Button"
 import { useCallback, useContext, useMemo } from "react"
-import { UserContext } from "../containers/UserProvider"
-import { Dropdown } from "./Navigation/Dropdown"
-import { Avatar } from "./User/Avatar"
-import { getUserProfileLink } from "../utils/user"
+import { UserContext } from "../../containers/UserProvider"
+import { Dropdown } from "../Navigation/Dropdown"
+import { Avatar } from "../User/Avatar"
+import { getUserProfileLink } from "../../utils/user"
 import { useState, useEffect } from "react"
 import { useRouter } from "next/router"
-import { SettingsModal } from "../containers/Settings/SettingsModal"
-import { SearchInputControlled } from "./Input/SearchInputControlled"
+import { SettingsModal } from "../../containers/Settings/SettingsModal"
+import { SearchInputControlled } from "../Input/SearchInputControlled"
 
 interface NavigationProps {
   onChangeSearchVisibility: (isVisible: boolean) => void
